fix(fabric): address review feedback on PR 19468
- build_workspace_container now derives qualified_name via make_workspace_name
- remove redundant _create_workspace_container wrapper in data_factory source
- move test_workspace_container_includes_external_url to fabric_common tests
- remove stale Workspace API docstring reference in onelake models
