+++ metadata-ingestion/src/datahub/ingestion/source/fabric/data_factory/source.py
@@ -262,7 +262,11 @@ class FabricDataFactorySource(StatefulIngestionSourceBase):
             # Pass 2: resolve edges and emit (cache is now fully populated)
             for workspace, pipeline_items in workspace_pipelines:
                 try:
-                    yield from self._create_workspace_container(workspace)
+                    yield from build_workspace_container(
+                        workspace=workspace,
+                        platform_instance=self.config.platform_instance,
+                        env=self.config.env,
+                    )
 
                     if self.config.extract_pipelines and pipeline_items:
                         invoke_pipeline_lineage = self._resolve_invoke_pipeline_edges(
@@ -344,15 +348,6 @@ class FabricDataFactorySource(StatefulIngestionSourceBase):
                 exc=e,
                 log=False,
             )
-
-    def _create_workspace_container(
-        self, workspace: FabricWorkspace
-    ) -> Iterable[Entity]:
-        yield from build_workspace_container(
-            workspace=workspace,
-            platform_instance=self.config.platform_instance,
-            env=self.config.env,
-        )
 
     @staticmethod
     def _get_pipeline_url(workspace_id: str, pipeline_id: str) -> str:
