+++ metadata-ingestion/tests/unit/fabric_common/__init__.py
@@ -1,0 +1,1 @@
+"""Unit tests for shared Fabric connector utilities."""
