+++ metadata-ingestion/tests/unit/fabric_onelake/test_source.py
@@ -5,10 +5,8 @@ from unittest.mock import MagicMock
 from datahub.emitter.mce_builder import datahub_guid
 from datahub.ingestion.source.fabric.common.models import (
     FABRIC_WORKSPACE_PLATFORM,
-    FabricWorkspace,
     WorkspaceKey,
 )
-from datahub.ingestion.source.fabric.common.utils import build_workspace_container
 from datahub.ingestion.source.fabric.onelake.source import (
     PLATFORM,
     FabricOneLakeSource,
@@ -144,26 +142,3 @@ def test_norm_respects_convert_urns_to_lowercase() -> None:
     src.config.convert_urns_to_lowercase = False
     assert FabricOneLakeSource._norm(src, "Sales") == "Sales"
     assert FabricOneLakeSource._norm(src, "CUSTOMERS") == "CUSTOMERS"
-
-
-def test_workspace_container_includes_external_url() -> None:
-    workspace = FabricWorkspace(
-        id="ws-123",
-        name="Test Workspace",
-        description="Test description",
-    )
-
-    container = next(
-        iter(
-            build_workspace_container(
-                workspace=workspace,
-                platform_instance="instance-1",
-                env="PROD",
-            )
-        )
-    )
-
-    assert (
-        container.external_url
-        == f"https://app.fabric.microsoft.com/groups/{workspace.id}/list"
-    )
